Vue/TS bug and bad typing fixes
