+++ frontend/src/components/widgets/inputs/RadioInput.vue
@@ -69,13 +69,15 @@
 <script lang="ts">
 import { defineComponent, PropType } from "vue";
 
+import { IconName } from "@/utilities/icons";
+
 import IconLabel from "@/components/widgets/labels/IconLabel.vue";
 import TextLabel from "@/components/widgets/labels/TextLabel.vue";
 
 export interface RadioEntryData {
 	value?: string;
 	label?: string;
-	icon?: string;
+	icon?: IconName;
 	tooltip?: string;
 	action?: () => void;
 }
